+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -139,6 +139,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added more custom pass API functions, mainly to render objects from another camera.
 - Added support for transparent Unlit in path tracing.
 - Added a minimal lit used for RTGI in peformance mode.
+- Added procedural metering mask that can follow an object
 
 ### Fixed
 - Fix when rescale probe all direction below zero (1219246)